Update1.0

+++ TowerDefense/Assets/Scripts/TowerStats.cs
@@ -6,11 +6,11 @@ using UnityEngine;
 public class TowerStats
 {
     public int level = 1;
-    public float attack = 1;
+    public float attack = 1f;
     public float fireRate = 1f;  
     public float range = 15f;
 
     [Header("Prices")]
-    public int priceUpgrade = 150;
-    public int priceSell = 80;
+    public float priceUpgrade = 150f;
+    public float priceSell = 80;
 }
